generate constants
