+++ client/package.json
@@ -7,7 +7,8 @@
     "build": "vue-cli-service build",
     "test": "vue-cli-service test:unit",
     "test:unit": "vue-cli-service test:unit",
-    "lint": "vue-cli-service lint"
+    "lint": "vue-cli-service lint",
+    "setup": "cd .. && sh generate_constants.sh"
   },
   "dependencies": {
     "@sentry/tracing": "^6.1.0",
