+++ generate_constants.sh
@@ -1,0 +1,5 @@
+cd server
+sh . ./run.sh --install -gc
+cd ..
+mkdir -p ./client/generated
+mv server/backend_constants.json ./client/generated/backend_constants.json
