+++ server/tests/test_halomod.py
@@ -109,13 +109,6 @@ def test_get_plot_data(client):
     assert "TheModel" in response.json["plot_data"]
 
 
-def test_constants(client):
-    response = client.get('/constants')
-    assert response is not None
-    assert response.status_code == 200
-    assert "cosmo_defaults" in response.json
-
-
 def test_create(client, create_payload):
     with client.session_transaction() as sess:
         sess["models"] = pickle.dumps({"TheModel": TracerHaloModel()})
